+++ src/pages/about.js
@@ -1,4 +1,5 @@
 import React from "react";
+
 const AboutPage = () => (
 	<div>
 		<h1> About australia.gov.au</h1>
@@ -9,15 +10,75 @@ const AboutPage = () => (
 		<br />
 		<h2>For the public</h2>
 		The australia.gov.au website is primarily for the public and is your
-		trusted source of government information. The australia.gov.au domain is
-		also a memorable website address for government campaigns used in the
-		media such as australia.gov.au/moneysmart - external site. The Digital
-		Transformation Agency - external site is responsible for the development
-		and ongoing operation of this website. You'll find this website is also
-		a source of government information that is not consolidated elsewhere,
-		such as public holidays. Additionally this website highlights particular
-		areas of government online presence, such as promoting government social
-		media profiles, or government mobile applications "apps".
+		trusted source of government information.
+		<br />
+		The australia.gov.au domain is also a memorable website address for
+		government campaigns used in the media such as
+		<a href="http://australia.gov.au/moneysmart">
+			{" "}
+			australia.gov.au/moneysmart{" "}
+		</a>{" "}
+		- external site.
+		<br />
+		You'll find this website is also a source of government information that
+		is not consolidated elsewhere, such as{" "}
+		<a href="placeholder">public holidays.</a>
+		<br />
+		Additionally this website highlights particular areas of government
+		online presence, such as promoting government{" "}
+		<a href="https://www.australia.gov.au/news-and-social-media/social-media">
+			social media profiles
+		</a>, or{" "}
+		<a href="https://www.australia.gov.au/content/apps-landing-page">
+			government mobile applications "apps".
+		</a>
+		<br />
+		The <a href="http://dta.gov.au">Digital Transformation Agency </a> -
+		external site is responsible for the development and ongoing operation
+		of this website.
+		<br />
+		<h2>For government agencies</h2>
+		The australia.gov.au domain is a memorable link shortener for government
+		campaigns used in the media such as australia.gov.au/moneysmart -
+		external site. Agencies don't need to register unnecessary domains for
+		content which can exist on their own sites.
+		<br />
+		Additionally, australia.gov.au highlights particular areas of government
+		online presence, such as promoting government{" "}
+		<a href="https://www.australia.gov.au/news-and-social-media/social-media">
+			social media profiles
+		</a>, or{" "}
+		<a href="https://www.australia.gov.au/content/apps-landing-page">
+			government mobile applications "apps".
+		</a>
+		<br />
+		Refer to the{" "}
+		<a href="https://www.australia.gov.au/terms-of-use">
+			Terms of use
+		</a>{" "}
+		page for the linking policy on australia.gov.au. Australian government
+		agencies can use the{" "}
+		<a href="https://www.australia.gov.au/content/help-and-contact-content-page">
+			contact{" "}
+		</a>us form to tell us about new sites or pages for inclusion in our
+		indexes.
+		<br />
+		australia.gov.au has been developed with usability, privacy and
+		accessibility principles in mind. We're always keen to improve the site.
+		If you have any feedback, ideas or even complaints, please send them in
+		writing to the following postal address:
+		<address>
+			<br />
+			australia.gov.au, Service Manager
+			<br />
+			Digital Transformation Agency
+			<br />
+			PO Box 457
+			<br />
+			Canberra ACT 2600
+			<br />
+			Australia
+		</address>
 	</div>
 );