+++ CHANGES.md
@@ -12,6 +12,7 @@
 -   Imported relevant govau assets
 -   Removed default styling generated from gatsby
 -   Installed `gatsby-plugin-favicon` to setup favicon
+-   Updated `About page` with more information and hyperlinks
 -   Added `linkComponent` to footer navigation links to stop page refresh
 -   Added `app-container` div to house web app with relevant padding for content
 -   Set footer to display `flex` to keep it sticky.